Adds regex testing for my Obsidian setup
tl;dr to test regexes that are used in the Obsidian plugin.

Wait, what?

In my journey to find a way to publish markdown notes directly from Obsidian,
I stumbled across an Obsidian plugin called obsidian-github-publisher. It
takes the markdown files and pushes them to a Github repo, converting
markdown links to Jekyll links in the process.

However, it is not built to respect the structure for images that I use in my
blog. To fix that, I have to enter regex into the plugin that then is used to
normalize the final artifact and leave in a format Jekyll understands.

I wrote about it in https://alexoliveira.cc/guide/jekyll-with-obsidian

The challenge is that the regex got complex and I can't test it quickly
anymore. To avoid relying on publishing and letting it error in github-pages,
or wait a minute or two to see the final result in Github, this file contains
those regexes with inputs and expected outputs. I can later just copy paste
them to the plugin and see if they work.

This site is useful for testing regexes: https://regex101.com/